refactor(content-system): move test stubs to src/Core/Test/Stub
Relocates ContentSystem test stubs from tests/_helper/ to src/Core/Test/Stub/ContentSystem/ so they are included in the installed Composer package and reusable by third-party plugins.

Converts @internal/final to @final per project stub convention and removes dead EnterTrackingVisitor.
