+++ src/Core/Test/Stub/ContentSystem/StubLoaderConfig.php
@@ -1,0 +1,14 @@
+<?php declare(strict_types=1);
+
+namespace Shopware\Core\Test\Stub\ContentSystem;
+
+use Shopware\Core\Framework\ContentSystem\Hydration\DataLoader\AbstractContentDataLoaderConfig;
+use Shopware\Core\Framework\Log\Package;
+
+/**
+ * @final
+ */
+#[Package('framework')]
+readonly class StubLoaderConfig extends AbstractContentDataLoaderConfig
+{
+}
