+++ tests/unit/Core/Framework/ContentSystem/Hydration/DataLoader/CurrencyLoader/CurrencyLoaderConfigSerializerTest.php
@@ -10,7 +10,7 @@ use Shopware\Core\Framework\ContentSystem\ContentSystemException;
 use Shopware\Core\Framework\ContentSystem\Hydration\DataLoader\CurrencyLoader\CurrencyLoaderConfig;
 use Shopware\Core\Framework\ContentSystem\Hydration\DataLoader\CurrencyLoader\CurrencyLoaderConfigSerializer;
 use Shopware\Core\Framework\Plugin\Exception\DecorationPatternException;
-use Shopware\Tests\Unit\Core\Framework\ContentSystem\_helper\StubLoaderConfig;
+use Shopware\Core\Test\Stub\ContentSystem\StubLoaderConfig;
 
 /**
  * @internal
