+++ tests/unit/Core/Framework/ContentSystem/Hydration/DataLoader/NavigationLoader/NavigationLoaderConfigSerializerTest.php
@@ -10,7 +10,7 @@ use Shopware\Core\Framework\ContentSystem\ContentSystemException;
 use Shopware\Core\Framework\ContentSystem\Hydration\DataLoader\NavigationLoader\NavigationLoaderConfig;
 use Shopware\Core\Framework\ContentSystem\Hydration\DataLoader\NavigationLoader\NavigationLoaderConfigSerializer;
 use Shopware\Core\Framework\Plugin\Exception\DecorationPatternException;
-use Shopware\Tests\Unit\Core\Framework\ContentSystem\_helper\StubLoaderConfig;
+use Shopware\Core\Test\Stub\ContentSystem\StubLoaderConfig;
 
 /**
  * @internal
