+++ tests/unit/Core/Framework/ContentSystem/Layout/Element/ContentElementTest.php
@@ -11,10 +11,10 @@ use Shopware\Core\Framework\ContentSystem\Hydration\DataContext\ContextType;
 use Shopware\Core\Framework\ContentSystem\Layout\Element\ContentElement;
 use Shopware\Core\Framework\ContentSystem\PlaceholderValues;
 use Shopware\Core\Framework\ContentSystem\RenderingSpecification;
-use Shopware\Tests\Unit\Core\Framework\ContentSystem\_helper\ContentElementBuilder;
-use Shopware\Tests\Unit\Core\Framework\ContentSystem\_helper\OrderTrackingVisitor;
-use Shopware\Tests\Unit\Core\Framework\ContentSystem\_helper\StubLoaderConfig;
-use Shopware\Tests\Unit\Core\Framework\ContentSystem\_helper\StubStruct;
+use Shopware\Core\Test\Stub\ContentSystem\ContentElementBuilder;
+use Shopware\Core\Test\Stub\ContentSystem\OrderTrackingVisitor;
+use Shopware\Core\Test\Stub\ContentSystem\StubLoaderConfig;
+use Shopware\Core\Test\Stub\ContentSystem\StubStruct;
 use Symfony\Component\HttpFoundation\Request;
 
 /**
