+++ tests/unit/Core/Framework/ContentSystem/Output/Format/SkeletonResponseFactoryTest.php
@@ -9,7 +9,7 @@ use Shopware\Core\Framework\ContentSystem\Output\Format\SkeletonResponseFactory;
 use Shopware\Core\Framework\ContentSystem\Output\Struct\ContentPage;
 use Shopware\Core\Framework\ContentSystem\RenderingMode;
 use Shopware\Core\Framework\ContentSystem\SalesChannel\ContentSkeletonRouteResponse;
-use Shopware\Tests\Unit\Core\Framework\ContentSystem\_helper\ContentElementBuilder;
+use Shopware\Core\Test\Stub\ContentSystem\ContentElementBuilder;
 
 /**
  * @internal
