--- tests/unit/Core/Framework/ContentSystem/_helper/OrderTrackingVisitor.php
@@ -1,29 +1,0 @@
-<?php declare(strict_types=1);
-
-namespace Shopware\Tests\Unit\Core\Framework\ContentSystem\_helper;
-
-use Shopware\Core\Framework\ContentSystem\Layout\Element\ContentElement;
-use Shopware\Core\Framework\ContentSystem\Layout\Element\Visitor\ElementVisitor;
-use Shopware\Core\Framework\Log\Package;
-
-/**
- * @internal
- */
-#[Package('framework')]
-final class OrderTrackingVisitor implements ElementVisitor
-{
-    /**
-     * @var list<string>
-     */
-    public array $log = [];
-
-    public function enter(ContentElement $element): void
-    {
-        $this->log[] = 'enter:' . $element->getComponent();
-    }
-
-    public function leave(ContentElement $element): void
-    {
-        $this->log[] = 'leave:' . $element->getComponent();
-    }
-}
