--- tests/unit/Core/Framework/ContentSystem/_helper/StubLoaderConfig.php
@@ -1,12 +1,0 @@
-<?php declare(strict_types=1);
-
-namespace Shopware\Tests\Unit\Core\Framework\ContentSystem\_helper;
-
-use Shopware\Core\Framework\ContentSystem\Hydration\DataLoader\AbstractContentDataLoaderConfig;
-
-/**
- * @internal
- */
-final readonly class StubLoaderConfig extends AbstractContentDataLoaderConfig
-{
-}
